story-to-mp3-audiobook-generator/requirements.txt
google-cloud-text-to-speech
pytest
pydub
numpy